% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_param.est.R
\name{estimate_lasso_parameters}
\alias{estimate_lasso_parameters}
\title{Estimate Lasso regression parameters}
\usage{
estimate_lasso_parameters(
  predictor,
  p_star_hat,
  gamma,
  penalty,
  prior,
  adaptive,
  sd_multiplier,
  elastic_net = F
)
}
\arguments{
\item{predictor}{A signature matrix}

\item{p_star_hat}{The regression target}

\item{gamma}{Hyperparameter for adaptive LASSO}

\item{penalty}{A vector of relative penalties on signatures}

\item{prior}{A vector of relative penalties on signatures given as prior}

\item{adaptive}{Boolean variable, perform adaptive LASSO or not}

\item{sd_multiplier}{How many SD should be tolerated from min MSE}

\item{elastic_net}{Use Elastic_net instead of LASSO?}
}
\value{
Returns a list of updated parameters including lambda picked, 
predictors used, a sequence of lambdas searched, alpha (elastic net) 
picked and the OLS fitting results (for adpative LASSO)
}
\description{
Determine the lambda in lasso procedure emperically
}
