% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_sigs}
\alias{plot_sigs}
\title{Visualize the output of siglasso as barplots}
\usage{
plot_sigs(sig_weights, plot_colors = NA, re_order = F, sample_order)
}
\arguments{
\item{sig_weights}{A matrix of weigths of signatures in multiple 
samples}

\item{plot_colors}{Colors for plotting, if NA, it generate its own colors}

\item{re_order}{Should it reorder the samples, default is False and 
always reorder if sample_order is provided}

\item{sample_order}{A vector of how samples should be ordered. If not 
supplied, it will order by the sum of weights}
}
\description{
Show a barplot of the signature fractions in multiple samples
}
