% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Visualize mutational spectrums of a sample, or all samples summarized}
\usage{
plot_spectrum(sig_weights, stat = "mean", std96sigs = T)
}
\arguments{
\item{sig_weights}{A matrix of mutational counts of different contexts}

\item{stat}{How to aggregate the spectrums. Choose "mean" or 
"median". Default is mean.}

\item{std96sigs}{Boolean variable, indicating if using 96
contexts and sorted by the contexts (from AC>AA to TT>GT)}
}
\description{
Show a plot of mutational spectrums of all samples. That is a plot of 
mutation counts catalogued by mutational contexts.
}
