\name{sigLASSO-package}
\alias{sigLASSO-package}
\alias{sigLASSO}
\docType{package}
\title{
\packageTitle{sigLASSO}
}
\description{
\packageDescription{sigLASSO}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sigLASSO}
\packageIndices{sigLASSO}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sigLASSO}

Maintainer: \packageMaintainer{sigLASSO}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
