% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siglasso.R
\name{siglasso}
\alias{siglasso}
\title{Attributing signatures using sigLASSO}
\usage{
siglasso(
  sample_spectrum,
  signature,
  conf = 0.1,
  prior,
  adaptive = T,
  gamma = 1,
  alpha_min = 400,
  iter_max = Inf,
  sd_multiplier = 0.5,
  elastic_net = F,
  plot = T
)
}
\arguments{
\item{sample_spectrum}{A data.frame/matrix of mutation counts 
catelogued by context}

\item{signature}{A signature matrix, in the form of possibilities 
of mutations. If not supplied, it loads 30 COSMIC signatures}

\item{conf}{Confidence factor of the mutational fitting (default: 0.1)}

\item{prior}{A vector of vector of weights on different signatures. 
(default: a vector of 1s (flat prior)).}

\item{adaptive}{A boolean variable indicating whether adaptive lasso should 
be used}

\item{gamma}{The hyperparameter gamma in adapative lasso. We find it is not 
sensitive to the analysis (default: 1)}

\item{alpha_min}{The  minimal alpha value (default:400)}

\item{iter_max}{Maximum iterations (default: Inf)}

\item{sd_multiplier}{cut-off for lambda tuning; how many sd it tolerates 
from min MSE.}

\item{elastic_net}{oolean variable indicating whether elastic net should 
be used}

\item{plot}{Boolean variable; if it should barplot the results}
}
\value{
A data.frame of weights of all signatures
}
\description{
Given a mutation spectrum of one or more sample, 
assign signatures with weights.
}
