% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siglasso_internal.R
\name{siglasso_internal}
\alias{siglasso_internal}
\title{Internal function for sigLASSO}
\usage{
siglasso_internal(
  spectrum,
  sig,
  prior,
  adaptive,
  elastic_net,
  gamma,
  alpha_min,
  iter_max,
  sd_multiplier,
  conf
)
}
\arguments{
\item{spectrum}{The mutation spectrum of the target sample}

\item{sig}{The signature matrix}

\item{prior}{The weights of penalties given as prior}

\item{adaptive}{Should it do adpative LASSO?}

\item{elastic_net}{Should it do elastic net?}

\item{gamma}{Hyperparameter for adpative LASSO}

\item{alpha_min}{The minimum alpha}

\item{iter_max}{The maximum number of iterations}

\item{sd_multiplier}{The tolerance of SEs in lambda searching}

\item{conf}{A confidence factor for linear fitting}
}
\description{
Internal function that assigns weights of signatures to a given sample
}
