% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{vcf2spec}
\alias{vcf2spec}
\title{Construct the spectrum from a VCF file}
\usage{
vcf2spec(
  bedtools_path = "bedtools",
  vcf_meta,
  ref_genome,
  output_file,
  context_length = 1,
  overwrite = F
)
}
\arguments{
\item{bedtools_path}{Path of bedtool executable, e.g. "~/bedtools/bin/bedtool"}

\item{vcf_meta}{A comma(CSV)/space/tab(TAB)-delimited meta vcf file in the format of "vcf_path[comma, space or tab]sample_name",
Alternatively, you can just supply a list of paths to vcfs without sample_name.
It will then use vcf file names as sample name.:}

\item{ref_genome}{An fa file path of reference genome sequence}

\item{output_file}{An output file that contains the context matrix}

\item{context_length}{The length of extension of each side of the mutation. 
By default it is 3 (trinucleotide context).}

\item{overwrite}{Should it overwrite the output_file if existed}
}
\value{
Returns a matrix of mutation counts in different context, 
in each of the samples
}
\description{
The function converts a vcf file into a spectrum matrix which has
mutation counts catalogued by contexts. It uses the fact that DNA has 
two complementary strands and folds the A/G mutations into to T/C. 
 
It is a wrapper on bedtools(https://github.com/arq5x/bedtools2/releases).
}
