% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{context2spec}
\alias{context2spec}
\title{Construct the spectrum from a mutation file with context}
\usage{
context2spec(input_muts, plot = TRUE, normalize = "none")
}
\arguments{
\item{input_muts}{A matrix of three columns; the first one is the context, 
with the central nucleotide mutated. The second one has the alternative
nucleotide. The third column stores the names of the samples. If it only has
two columns, the function assumes all mutations are from one single sample}

\item{plot}{a binary indicator whethere a plot of the specturm will be made 
(default: TRUE)}

\item{normalize}{Normalization choice ("none", "genome", "exome"). In general,
when fitting COSMIC signatures, which are derived from a mixture of WGS and WES 
smaples, we do not think there is additional benefits with normalization (default: none)
Normalization uses the trinucleotides frequencies in the genome.
Please do not use normalized frequencies for futher fitting (siglasso()) as it 
the mutation counts are no longer discrete. siglass() has an option for such 
normalization, operating on the signatures.}
}
\value{
Returns a matrix of mutation counts in different context, 
in each of the samples
}
\description{
The function converts mutation with reference contexts 
into a spectrum, which has mutation counts catalogued by 
contexts. It uses the fact that DNA has two complementary strands
and folds the A/G mutations into to T/C.
}
