% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess (Autosaved).R, R/preprocess.R
\name{reverse_context}
\alias{reverse_context}
\title{Get the reverse complementary sequence}
\usage{
reverse_context(mut_context, alt_nuc)

reverse_context(mut_context, alt_nuc)
}
\arguments{
\item{mut_context}{A string of ref nucleotide context}

\item{alt_nuc}{The alternative nucleotide}

\item{mut_context}{A string of ref nucleotide context}

\item{alt_nuc}{The alternative nucleotide}
}
\value{
Returns a string of nucleotide context of the same length

Returns a string of nucleotide context of the same length
}
\description{
Given a mutational context, get the context on its complementary
strand

Given a mutational context, get the context on its complementary
strand
}
