% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{vcf2spec}
\alias{vcf2spec}
\title{Construct the spectrum from a vcf file}
\usage{
vcf2spec(vcf_file, ref_genome, context_length = 3)
}
\arguments{
\item{vcf_file}{A vcf file}

\item{ref_genome}{An fa file of reference genome sequence}

\item{context_length}{The length of context used. By default it is 3.}
}
\value{
Returns a matrix of mutation counts in different context, 
in each of the samples
}
\description{
The function converts a vcf file into a spectrum matrix which has
mutation counts catalogued by contexts. It uses the fact that DNA has 
two complementary strands and folds the A/G mutations into to T/C.
}
